fix RDB HiTrace by MACRO

+++ frameworks/native/dataability/src/ishared_result_set_proxy.cpp
@@ -62,8 +62,7 @@ int ISharedResultSetProxy::GetAllColumnNames(std::vector<std::string> &columnNam
         return E_INVALID_PARCEL;
     }
     columnNames_ = columnNames;
-    for (int i = 0; i < columnNames_.size(); i++)
-    {
+    for (size_t i = 0; i < columnNames_.size(); i++) {
         std::string columnNameStr = columnNames_[i];
         transform(columnNameStr.begin(), columnNameStr.end(), columnNameStr.begin(), ::tolower);
         columnMap_[columnNameStr] = i;
